cleanup

+++ minix/cpumeter.h
@@ -15,18 +15,17 @@
 
 class CPUMeter : public FieldMeterGraph {
 public:
-    CPUMeter(unsigned int cpu=0);
+    CPUMeter(size_t cpu=0);
 
     virtual std::string resName(void) const override { return "cpu"; }
     virtual void checkevent(void) override;
-
     virtual void checkResources(const ResDB &rdb) override;
 
     static size_t countCPUs(void);
     static std::string cpuStr(size_t num);
 
 private:
-    unsigned int _cpu;
+    size_t _cpu;
     std::map<pid_t, XOSVProc> _ptable;
 
     std::vector<uint64_t> getTicks(void);
